solve.c: fix int32 overflow in consolidation hash resize trigger
When checkpoint resume pre-sized the consolidation hash to 2^30 slots
(needed for >300M records to avoid clustering stall), the resize-trigger
expression `ht_size * 3 / 4` overflowed signed int. ht_size = 2^30,
ht_size * 3 = 3.2B > INT_MAX (2.15B), wrapped to negative. The
`solution_count > negative` comparison was always true, so resize fired
on every insert. resize_hash_table then tried to allocate 2^31 slots
(64 GB), which itself overflows int new_size = 1 << 31, calloc fails
or reports negative MB ("-65536 MB"), FATAL exit.

Surfaced when resuming the 100T pilot on pilot-100T-westus3 after a
spot eviction at 17:59 UTC 2026-04-28. Estimated 680M records on disk;
pre-size to 2^30 is correct but the spurious-resize path killed startup.

Fix: cast `ht_size * 3 / 4` to long long before multiply. One-line
change. Selftest sha 403f7202 still byte-identical.

+++ solve.c
@@ -2106,7 +2106,7 @@ static long long sub_ckpt_load(ThreadState *consolidate_into) {
                         fflush(stderr);
                         warns_this_file++;
                     }
-                    if (consolidate_into->solution_count > consolidate_into->ht_size * 3 / 4) {
+                    if (consolidate_into->solution_count > (long long)consolidate_into->ht_size * 3 / 4) {
                         resize_hash_table(consolidate_into);
                     }
                     break;
